third_party/WebKit/LayoutTests/fast/css-intrinsic-dimensions/block-min-width-and-max-width.html

<!DOCTYPE html>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<script src="../../resources/check-layout-th.js"></script>

<p>There should be a hotpink <em>square</em> below.</p>
<div id="container" style="display:flow-root;">
    <div data-expected-width="300" style="float:left; background:hotpink;">
        <div style="height:300px; min-width:300px; max-width:10px;"></div>
    </div>
</div>

<script>
    checkLayout("#container");
</script>
